P2P: Fix validation on Invitation Request error path
It was possible for the error path to try to use P2P Group ID attribute
even if one was not included in the message. This could result in
dereferencing a NULL pointer, so re-check the pointer before copying the
data.

CRs-Fixed: 704007
Git-commit: fa72a880ed34929ed801b89afb2e243d49e06e07
Git-repo : git://w1.fi/srv/git/hostap.git
Signed-off-by: Jouni Malinen <j@w1.fi>

Change-Id: I9ddc18c8248ece8ca9bb5d074a069f5129058a05