Validate driver extended capabilities length against buffer length
Prepare for new extended capabilities bits by checking that the local
buffer is large enough to contain all the bits the driver requests. The
existing buffers are large enough to include anything defined until now,
but it would be possible to add more definitions in the future, so
increase them a bit as well to make this more future proof.

CRs-Fixed: 704007
Git-commit: 0bbaa9b93f76025e81c3976bebcc00bf1660cfec
Git-repo : git://w1.fi/srv/git/hostap.git
Signed-off-by: Jouni Malinen <j@w1.fi>

Change-Id: Icc1a6e662896deac7d0506c5e8475eeedf554cb4