hostapd: Configure driver ACL even if MAC address list is empty
Earlier commit related to MAC address based access control list
offloaded to the driver was not sending ACL configuration to the driver
if the MAC address list was empty. Remove this check as empty access
control list is a valid use case and sending ACL parameters should not
be dependent on whether the list is empty.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: cf1600ace434de06a402f3eaa3661a94b3a6f20f
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-fixed: 662506
Change-Id: I2725184ea35100dc99f449366154d6389a2d76f0