Add support for driver command to update roaming policy
The network block bssid parameter can be used to force a specific BSS to
be used for a connection. It is also possible to modify this parameter
during an association. Previously, that did not result in any
notification to the driver which was somewhat problematic with drivers
that take care of BSS selection. Add a new mechanism to allow
wpa_supplicant to provide a driver update if the bssid parameter change
for the current connection modifies roaming policy (roaming
allowed/disallowed within ESS).

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Git-commit: 0ef023e47863ed92367fca9a65ace6f329eb3d8c
Git-repo : git://w1.fi/srv/git/hostap.git
CRs-fixed: 748007

Change-Id: Ia1f34a25d2e3e5f70f6d2cd90f6ef27b94ede068