P2P: Ignore remain-on-channel callback event if not waiting for one
It was possible for the previously requested remain-on-channel for P2P
listen state to get canceled before having received driver event
indicating start of that remain-on-channel operation. In such a case,
the event was able to trigger P2P module to start processing listen
timeout even though there was not supposed to be a following listen
operation anymore. Skip the driver event if we are not waiting for a new
listen state.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: f2dc06e2bd8116f0c940c24bc6c4a0f615573dcb
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: Ib1c3d0e4c3368a0afd5e35eb05ae2a9baf61c7dd
CRs-Fixed: 792354